chore: rename JWT to Jwt, make jwt package
