+++ src/main/java/com/mapu/infra/oauth/jwt/JwtUtil.java
@@ -1,4 +1,4 @@
-package com.mapu.infra.oauth;
+package com.mapu.infra.oauth.jwt;
 
 import io.jsonwebtoken.Jwts;
 import jakarta.servlet.http.Cookie;
@@ -11,11 +11,11 @@ import java.nio.charset.StandardCharsets;
 import java.util.Date;
 
 @Component
-public class JWTUtil {
+public class JwtUtil {
 
     private SecretKey secretKey;
 
-    public JWTUtil(@Value("${spring.jwt.secret}")String secret) {
+    public JwtUtil(@Value("${spring.jwt.secret}")String secret) {
 
         secretKey = new SecretKeySpec(secret.getBytes(StandardCharsets.UTF_8), Jwts.SIG.HS256.key().build().getAlgorithm());
     }
@@ -50,7 +50,7 @@ public class JWTUtil {
 
         Cookie cookie = new Cookie(key, value);
         cookie.setMaxAge(60*60*60);
-        //cookie.setSecure(true);
+        //cookie.setSecure(true); //TODO https로 전환 후에 주석 해제
         cookie.setPath("/");
         cookie.setHttpOnly(true);
